requirements.txt
git+https://github.com/xivo-pbx/xivo-lib-python.git
pyyaml==3.10
